Remove attributes that open recipe links in new page
After adding a home page link at the top of each recipe's webpage,
it didn't quite make sense for me to default to having each recipe
link on the home page opening in a new window when clicked. Since
the recipes aren't "external" links going to another website, it makes
more sense to keep the "internal" logic consistent by having links open
in the same window; I reckon that with this change, users will not have
the feeling that they are being taken "elsewhere" on the web.